refactor(ui): relocate terms of use modal
relocate terms of use modal
